+++ src/test/java/models/ProjectBuilder.java
@@ -1,7 +1,6 @@
 package models;
 
 import com.google.gson.annotations.SerializedName;
-import enums.ProjectType;
 import lombok.Builder;
 import lombok.Data;
