+++ src/test/java/steps/MilestoneSteps.java
@@ -1,17 +1,17 @@
-package steps;
-
-import baseEntities.BaseStep;
-import models.Milestone;
-import models.Project;
-import org.openqa.selenium.WebDriver;
-
-public class MilestoneSteps extends BaseStep {
-
-    public MilestoneSteps(WebDriver driver) {
-        super(driver);
-    }
-
-    public void createMilestone(Project project, Milestone milestone) {
-
-    }
-}
+//package steps;
+//
+//import baseEntities.BaseStep;
+//import models.Milestone;
+//import models.Project;
+//import org.openqa.selenium.WebDriver;
+//
+//public class MilestoneSteps extends BaseStep {
+//
+//    public MilestoneSteps(WebDriver driver) {
+//        super(driver);
+//    }
+//
+//    public void createMilestone(Project project, Milestone milestone) {
+//
+//    }
+//}
