+++ src/test/java/tests/api/ReqresApiTest.java
@@ -1,59 +1,59 @@
-package tests.api;
-
-import io.restassured.RestAssured;
-import io.restassured.http.Method;
-import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
-import org.apache.http.HttpStatus;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import static io.restassured.RestAssured.given;
-
-public class ReqresApiTest {
-
-    @Test
-    public void simpleStepByStepApiTest(){
-        // Setup RestAssured
-        RestAssured.baseURI = "https://reqres.in";
-
-        // Setup endpoint
-        int userID = 2;
-        String endpoint = "/api/users/" + userID;
-
-        // Setup request Object
-        RequestSpecification httpRequest = given();
-
-        // Setup Response Object
-
-        Response response = httpRequest.request(Method.GET, endpoint);
-
-        //Get response Status
-        int statusCode = response.getStatusCode();
-        System.out.println("Status Code:" + statusCode);
-        Assert.assertEquals(statusCode, 200); //проверка 200
-        Assert.assertEquals(statusCode, HttpStatus.SC_OK); // такое же как первое
-
-        // Get Response body
-        String responseBody = response.getBody().asString();
-        System.out.println("Response: " + responseBody);
-
-    }
-    @Test
-    public void simpleShortApiTest(){
-        // Setup RestAssured
-        RestAssured.baseURI = "https://reqres.in";
-
-        // Setup endpoint
-        int userID = 2;
-        String endpoint = "/api/users/" + userID;
-
-        given()
-                .when()
-                .get(endpoint)
-                .then()
-                .statusCode(HttpStatus.SC_OK)
-                .log().body();
-
-    }
-}
+//package tests.api;
+//
+//import io.restassured.RestAssured;
+//import io.restassured.http.Method;
+//import io.restassured.response.Response;
+//import io.restassured.specification.RequestSpecification;
+//import org.apache.http.HttpStatus;
+//import org.testng.Assert;
+//import org.testng.annotations.Test;
+//
+//import static io.restassured.RestAssured.given;
+//
+//public class ReqresApiTest {
+//
+//    @Test
+//    public void simpleStepByStepApiTest(){
+//        // Setup RestAssured
+//        RestAssured.baseURI = "https://reqres.in";
+//
+//        // Setup endpoint
+//        int userID = 2;
+//        String endpoint = "/api/users/" + userID;
+//
+//        // Setup request Object
+//        RequestSpecification httpRequest = given();
+//
+//        // Setup Response Object
+//
+//        Response response = httpRequest.request(Method.GET, endpoint);
+//
+//        //Get response Status
+//        int statusCode = response.getStatusCode();
+//        System.out.println("Status Code:" + statusCode);
+//        Assert.assertEquals(statusCode, 200); //проверка 200
+//        Assert.assertEquals(statusCode, HttpStatus.SC_OK); // такое же как первое
+//
+//        // Get Response body
+//        String responseBody = response.getBody().asString();
+//        System.out.println("Response: " + responseBody);
+//
+//    }
+//    @Test
+//    public void simpleShortApiTest(){
+//        // Setup RestAssured
+//        RestAssured.baseURI = "https://reqres.in";
+//
+//        // Setup endpoint
+//        int userID = 2;
+//        String endpoint = "/api/users/" + userID;
+//
+//        given()
+//                .when()
+//                .get(endpoint)
+//                .then()
+//                .statusCode(HttpStatus.SC_OK)
+//                .log().body();
+//
+//    }
+//}
