+++ src/test/java/tests/api/TestRailApiTest.java
@@ -1,60 +1,60 @@
-package tests.api;
-
-import baseEntities.BaseApiTest;
-import com.tms.core.ReadProperties;
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.http.Method;
-import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
-import org.apache.http.HttpStatus;
-import org.apache.http.protocol.HTTP;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import static io.restassured.RestAssured.given;
-
-public class TestRailApiTest extends BaseApiTest {
-
-    @Test
-    public void getAllUsers() {
-        // Setup RestAssured
-        RestAssured.baseURI = ReadProperties.getUrl();
-
-        // Setup endpoint
-        String endpoint = "/index.php?/api/v2/get_users";
-
-        // Setup request Object
-        RequestSpecification httpRequest = given();
-        httpRequest.header(HTTP.CONTENT_TYPE, ContentType.JSON);
-        httpRequest.auth().preemptive().basic(ReadProperties.getUsername(), ReadProperties.getPassword());
-
-        // Setup Response Object
-
-        Response response = httpRequest.request(Method.GET, endpoint);
-
-        //Get response Status
-        int statusCode = response.getStatusCode();
-        System.out.println("Status Code:" + statusCode);
-        Assert.assertEquals(statusCode, 200); //проверка 200
-        Assert.assertEquals(statusCode, HttpStatus.SC_OK); // такое же как первое
-
-        // Get Response body
-        String responseBody = response.getBody().asString();
-        System.out.println("Response: " + responseBody);
-    }
-
-    @Test
-    public void getAllUsers1() {
-        // Setup endpoint
-        String endpoint = "/index.php?/api/v2/get_users";
-
-        given()
-                .when()
-                .get(endpoint)
-                .then()
-                .log().status()
-                .log().body()
-                .statusCode(HttpStatus.SC_OK);
-    }
-}
+//package tests.api;
+//
+//import baseEntities.BaseApiTest;
+//import core.ReadProperties;
+//import io.restassured.RestAssured;
+//import io.restassured.http.ContentType;
+//import io.restassured.http.Method;
+//import io.restassured.response.Response;
+//import io.restassured.specification.RequestSpecification;
+//import org.apache.http.HttpStatus;
+//import org.apache.http.protocol.HTTP;
+//import org.testng.Assert;
+//import org.testng.annotations.Test;
+//
+//import static io.restassured.RestAssured.given;
+//
+//public class TestRailApiTest extends BaseApiTest {
+//
+//    @Test
+//    public void getAllUsers() {
+//        // Setup RestAssured
+//        RestAssured.baseURI = ReadProperties.getUrl();
+//
+//        // Setup endpoint
+//        String endpoint = "/index.php?/api/v2/get_users";
+//
+//        // Setup request Object
+//        RequestSpecification httpRequest = given();
+//        httpRequest.header(HTTP.CONTENT_TYPE, ContentType.JSON);
+//        httpRequest.auth().preemptive().basic(ReadProperties.getUsername(), ReadProperties.getPassword());
+//
+//        // Setup Response Object
+//
+//        Response response = httpRequest.request(Method.GET, endpoint);
+//
+//        //Get response Status
+//        int statusCode = response.getStatusCode();
+//        System.out.println("Status Code:" + statusCode);
+//        Assert.assertEquals(statusCode, 200); //проверка 200
+//        Assert.assertEquals(statusCode, HttpStatus.SC_OK); // такое же как первое
+//
+//        // Get Response body
+//        String responseBody = response.getBody().asString();
+//        System.out.println("Response: " + responseBody);
+//    }
+//
+//    @Test
+//    public void getAllUsers1() {
+//        // Setup endpoint
+//        String endpoint = "/index.php?/api/v2/get_users";
+//
+//        given()
+//                .when()
+//                .get(endpoint)
+//                .then()
+//                .log().status()
+//                .log().body()
+//                .statusCode(HttpStatus.SC_OK);
+//    }
+//}
