+++ src/test/java/utils/Listener.java
@@ -1,7 +1,7 @@
 package utils;
 
-        import org.testng.ITestListener;
-        import org.testng.ITestResult;
+import org.testng.ITestListener;
+import org.testng.ITestResult;
 
 public class Listener implements ITestListener {
     public void onTestFailure(ITestResult result) {
