+++ src/test/java/utils/Waits.java
@@ -1,6 +1,6 @@
 package utils;
 
-import com.tms.core.ReadProperties;
+import core.ReadProperties;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
